Moved some context-related classes out of the plan visitors package and reduced visibility
