+++ integration/spark/src/main/java/io/openlineage/spark/agent/lifecycle/ContextFactory.java
@@ -5,10 +5,9 @@ import io.openlineage.spark.agent.OpenLineageContext;
 import io.openlineage.spark.agent.lifecycle.plan.BigQueryNodeVisitor;
 import io.openlineage.spark.agent.lifecycle.plan.CommandPlanVisitor;
 import io.openlineage.spark.agent.lifecycle.plan.DatasetSourceVisitor;
-import io.openlineage.spark.agent.lifecycle.plan.InputDatasetVisitors;
 import io.openlineage.spark.agent.lifecycle.plan.LogicalRDDVisitor;
 import io.openlineage.spark.agent.lifecycle.plan.LogicalRelationVisitor;
-import io.openlineage.spark.agent.lifecycle.plan.OutputDatasetVisitors;
+
 import java.util.ArrayList;
 import java.util.List;
 import lombok.AllArgsConstructor;
